Fixing bug in old migration (so it can be executed on empty database). Updating spec file (fixing access rights).

+++ recodex-api.spec
@@ -2,8 +2,8 @@
 %define short_name api
 %define install_dir /opt/%{name}
 %define version 1.22.0
-%define unmangled_version 9d6d8a06f6030364d5c2490fd723bf19c015f9ed
-%define release 7
+%define unmangled_version 288449c5bf48e818735a55bef98756c05e31e961
+%define release 12
 
 Summary: ReCodEx core API component
 Name: %{name}
@@ -59,6 +59,16 @@ php %{install_dir}/composer-stable.phar install --no-ansi --no-dev --no-interact
 # Run cleaner after installation
 %{install_dir}/cleaner
 
+setfacl -Rd -m u:apache:rwX %{install_dir}/temp
+setfacl -Rd -m u:recodex:rwX %{install_dir}/temp
+setfacl -Rd -m u:apache:rwX %{install_dir}/job_config
+setfacl -Rd -m u:recodex:rwX %{install_dir}/job_config
+setfacl -Rd -m u:apache:rwX %{install_dir}/uploaded_data
+setfacl -Rd -m u:recodex:rwX %{install_dir}/uploaded_data
+setfacl -Rd -m u:apache:rwX /var/log/recodex/core-api
+setfacl -Rd -m u:recodex:rwX /var/log/recodex/core-api
+
+
 %postun
 
 %pre
@@ -67,18 +77,20 @@ getent passwd recodex >/dev/null || useradd -r -g recodex -d %{_sysconfdir}/reco
 exit 0
 
 %preun
-%{install_dir}/cleaner
-rm -rf %{install_dir}/vendor
+if [ $1 == 0 ]; then
+	%{install_dir}/cleaner
+	rm -rf %{install_dir}/vendor
+fi
 
 %files
 %defattr(-,recodex,recodex)
 %dir %{install_dir}
 %dir %attr(-,recodex,recodex) %{_sysconfdir}/recodex/core-api
-%dir %attr(0755,apache,recodex) /var/log/recodex/core-api
-%attr(0755,apache,recodex) %{install_dir}/log
-%dir %attr(0755,apache,recodex) %{install_dir}/job_config
-%dir %attr(0755,apache,recodex) %{install_dir}/uploaded_data
-%dir %attr(0755,apache,recodex) %{install_dir}/temp
+%dir %attr(0775,apache,recodex) /var/log/recodex/core-api
+%attr(0775,apache,recodex) %{install_dir}/log
+%dir %attr(0775,apache,recodex) %{install_dir}/job_config
+%dir %attr(0775,apache,recodex) %{install_dir}/uploaded_data
+%dir %attr(0775,apache,recodex) %{install_dir}/temp
 %dir %{install_dir}/app
 
 %{install_dir}/app/bootstrap.php
@@ -95,7 +107,7 @@ rm -rf %{install_dir}/vendor
 %{install_dir}/composer.json
 %{install_dir}/composer.lock
 %{install_dir}/composer-stable.phar
-%attr(0755,recodex,recodex) %{install_dir}/cleaner
+%attr(0770,recodex,recodex) %{install_dir}/cleaner
 %{install_dir}/migrations
 %{install_dir}/www
 %attr(0660,apache,recodex) %{_sysconfdir}/recodex/core-api/config.local.neon
